requirements.txt
torch
torchsde
einops
transformers>=4.25.1
safetensors>=0.3.0
aiohttp
accelerate
pyyaml
Pillow
scipy
tqdm
psutil
requests
opencv-python
numba
clip-interrogator
simpleeval
opencv-python
insightface
uuid
onnxruntime
get-port